reorganzied a lot of code and broke some code

+++ src/com/devon/infiniteworld/entities/Entity.java
@@ -15,7 +15,7 @@ public class Entity
 		this.position = new Vector2f(position.x, position.y);
 	}
 	
-	public void draw(int x, int y)
+	public void draw(float x, float y)
 	{
 		this.image.draw(x, y);
 	}
